Move TestHelper.race tests into @Isolated classes: Flowable flatMap/mergeWith family (#8075)
